// examples/cloudflare-neon-prisma/src/prisma/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Post: {
      readonly id: Char<36>;
      readonly title: CodecTypes["pg/text@1"]["output"];
      readonly authorId: Char<36>;
    };
    readonly User: {
      readonly id: Char<36>;
      readonly email: CodecTypes["pg/text@1"]["output"];
      readonly name: CodecTypes["pg/text@1"]["output"];
      readonly createdAt: CodecTypes["pg/timestamptz-string@1"]["output"];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Post: {
      readonly id: CodecTypes["sql/char@1"]["input"];
      readonly title: CodecTypes["pg/text@1"]["input"];
      readonly authorId: CodecTypes["sql/char@1"]["input"];
    };
    readonly User: {
      readonly id: CodecTypes["sql/char@1"]["input"];
      readonly email: CodecTypes["pg/text@1"]["input"];
      readonly name: CodecTypes["pg/text@1"]["input"];
      readonly createdAt: CodecTypes["pg/timestamptz-string@1"]["input"];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly Post: {
      readonly authorId: Char<36>;
      readonly id: Char<36>;
      readonly title: CodecTypes["pg/text@1"]["output"];
    };
    readonly User: {
      readonly createdAt: CodecTypes["pg/timestamptz-string@1"]["output"];
      readonly email: CodecTypes["pg/text@1"]["output"];
      readonly id: Char<36>;
      readonly name: CodecTypes["pg/text@1"]["output"];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly Post: {
      readonly authorId: CodecTypes["sql/char@1"]["input"];
      readonly id: CodecTypes["sql/char@1"]["input"];
      readonly title: CodecTypes["pg/text@1"]["input"];
    };
    readonly User: {
      readonly createdAt: CodecTypes["pg/timestamptz-string@1"]["input"];
      readonly email: CodecTypes["pg/text@1"]["input"];
      readonly id: CodecTypes["sql/char@1"]["input"];
      readonly name: CodecTypes["pg/text@1"]["input"];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;